fix: 피드백: [각 역할에 맞는 분리 및 입력 값에 대한 기본적인 가공(https://github.com/next-step/java-ladder/pull/859#discussion_r609644819)
