--- src/main/java/nextstep/ladder/controller/LadderParameterHelper.java
@@ -1,15 +1,0 @@
-package nextstep.ladder.controller;
-
-import java.util.function.Supplier;
-
-public final class LadderParameterHelper {
-    private LadderParameterHelper() {}
-
-    static <T> T parseArgumentResolver(final Supplier<T> method) {
-        try {
-            return method.get();
-        } catch (RuntimeException e) {
-            throw new IllegalArgumentException();
-        }
-    }
-}
